requirements.txt
cv2imshow==0.10
keras==3.5.0
matplotlib==3.9.2
opencv-python==4.10.0.84
tensorflow==2.17.0